pa-workflow/ocr-service/requirements.txt
flask
flask-cors
paddlepaddle
paddleocr
pdf2image
python-docx
Pillow
openpyxl
python-pptx
numpy
pypdf
